Frontend/[final]django/requirements.txt
Django==3.2.8
django-allauth==0.50.0
django-etc==1.3.1
django-extensions==3.1.5
django-hitcount==1.3.5
django-markdownx==3.0.1
django-shell-ipynb==0.3.0
Markdown==3.3.4
gunicorn==20.1.0
django-defender==0.9.1
pandas
torch
psycopg2-binary==2.9.1
Pillow==8.4.0
urllib3==1.26.7

appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1

django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
idna==3.3
ipython==7.29.0
jedi==0.18.0
keyboard==0.13.5
MouseInfo==0.1.3
oauthlib==3.1.1
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python3-openid==3.2.0
pytweening==1.0.4
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rubicon-objc==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5 
webencodings==0.5.1
tqdm